PIZDEC

+++ figureCollection.cpp
@@ -13,7 +13,7 @@ void FigureCollection::removeByIndex(std::size_t index) {
     if (index >= figures.size()) {
         throw std::out_of_range("\nError: index out of range.\n");
     }
-    figures.erase(figures.begin() + static_cast<long>(index));
+    figures.erase(figures.begin() + index);
 }
 
 void FigureCollection::removeWithPerimeterGreaterThan(double limit) {
